Fix an issue with files smaller than chunk size.
If chunking.mandatory is set to true in the client but the file is smaller than the chunk size, the file is uploaded directly to the 'files' folder bypassing the 'chunks' folder. The call

$result = $uploader->combineChunks("files");

then causes the correctly uploaded file to be overwritten by an empty file, since the necessary chunks do not exist.